requirements.txt
uvicorn
fastapi[all]
pydantic
openai
langchain
langchain_openai
pypdf
chromadb
python-dotenv

## PII
spacy
pydantic
spacy_fastlang
presidio-analyzer
presidio-anonymizer
faker